[dynamo] Maintainable code - Move decorators in a separate file
[ghstack-poisoned]
