+++ torch/_higher_order_ops/wrap.py
@@ -14,7 +14,7 @@ class Wrap(HigherOrderOperator):
         # Dynamo already traces the body of HigherOrderOp beforehand when it
         # so no need to trace into it.
         import torch._dynamo  # noqa: F401
-        from torch._dynamo.eval_frame import disable
+        from torch._dynamo import disable
 
         @disable
         def wrapper():
